Eliminate specialization for Tensor<1,dim>
This is more or less an exact code copy of Tensor<rank_,dim>.

tensor_base.h now contains the full definition of Tensor<rank_,dim>. All
external functions operating on tensors that are not strictly needed reside
in tensor.h.
