+++ include/deal.II/base/tensor.h
@@ -19,703 +19,11 @@
 
 #include <deal.II/base/config.h>
 #include <deal.II/base/tensor_base.h>
-#include <deal.II/base/utilities.h>
 
 DEAL_II_NAMESPACE_OPEN
-template <int rank_, int dim, typename Number> class Tensor;
-template <int dim, typename Number> class Tensor<1,dim,Number>;
 
-/**
- * A general tensor class with an arbitrary rank, i.e. with an arbitrary
- * number of indices. The Tensor class provides an indexing operator and a bit
- * of infrastructure, but most functionality is recursively handed down to
- * tensors of rank 1 or put into external templated functions, e.g. the
- * <tt>contract</tt> family.
- *
- * Using this tensor class for objects of rank 2 has advantages over matrices
- * in many cases since the dimension is known to the compiler as well as the
- * location of the data. It is therefore possible to produce far more
- * efficient code than for matrices with runtime-dependent dimension. It also
- * makes the code easier to read because of the semantic difference between a
- * tensor (an object that relates to a coordinate system and has
- * transformation properties with regard to coordinate rotations and
- * transforms) and matrices (which we consider as operators on arbitrary
- * vector spaces related to linear algebra things).
- *
- * @tparam rank_ An integer that denotes the rank of this tensor. A rank-0
- * tensor is a scalar, a rank-1 tensor is a vector with @p dim components, a
- * rank-2 tensor is a matrix with dim-by-dim components, etc. There are
- * specializations of this class for rank-0 and rank-1 tensors. There is also
- * a related class SymmetricTensor for tensors of even rank whose elements are
- * symmetric.
- * @tparam dim An integer that denotes the dimension of the space in which
- * this tensor operates. This of course equals the number of coordinates that
- * identify a point and rank-1 tensor.
- * @tparam Number The data type in which the tensor elements are to be stored.
- * This will, in almost all cases, simply be the default @p double, but there
- * are cases where one may want to store elements in a different (and always
- * scalar) type. It can be used to base tensors on @p float or @p complex
- * numbers or any other data type that implements basic arithmetic operations.
- * Another example would be a type that allows for Automatic Differentiation
- * (see, for example, the Sacado type used in step-33) and thereby can
- * generate analytic (spatial) derivatives of a function that takes a tensor
- * as argument.
- *
- * @ingroup geomprimitives
- * @author Wolfgang Bangerth, 1998-2005
- */
-template <int rank_, int dim, typename Number>
-class Tensor
-{
-public:
-  /**
-   * Provide a way to get the dimension of an object without explicit
-   * knowledge of it's data type. Implementation is this way instead of
-   * providing a function <tt>dimension()</tt> because now it is possible to
-   * get the dimension at compile time without the expansion and preevaluation
-   * of an inlined function; the compiler may therefore produce more efficient
-   * code and you may use this value to declare other data types.
-   */
-  static const unsigned int dimension = dim;
 
-  /**
-   * Publish the rank of this tensor to the outside world.
-   */
-  static const unsigned int rank = rank_;
 
-  /**
-   * Number of independent components of a tensor of current rank. This is dim
-   * times the number of independent components of each sub-tensor.
-   */
-  static const unsigned int
-  n_independent_components = Tensor<rank_-1,dim>::n_independent_components *dim;
-
-  /**
-   * Type of stored objects. This is a tensor of lower rank.
-   */
-  typedef Tensor<rank_-1,dim,Number> value_type;
-
-  /**
-   * Declare a type that has holds real-valued numbers with the same precision
-   * as the template argument to this class. For std::complex<number>, this
-   * corresponds to type number, and it is equal to Number for all other
-   * cases. See also the respective field in Vector<Number>.
-   *
-   * This typedef is used to represent the return type of norms.
-   */
-  typedef typename numbers::NumberTraits<Number>::real_type real_type;
-
-  /**
-   * Declare an array type which can be used to initialize an object of this
-   * type statically.
-   */
-  typedef typename Tensor<rank_-1,dim,Number>::array_type array_type[dim];
-
-  /**
-   * Constructor. Initialize all entries to zero if
-   * <tt>initialize==true</tt>; this is the default behaviour.
-   */
-  explicit
-  Tensor (const bool initialize = true);
-
-  /**
-   * Copy constructor.
-   */
-  Tensor (const Tensor<rank_,dim,Number> &initializer);
-
-  /**
-   * Constructor, where the data is copied from a C-style array.
-   */
-  Tensor (const array_type &initializer);
-
-  /**
-   * Constructor from tensors with different underlying scalar type. This
-   * obviously requires that the @p OtherNumber type is convertible to @p
-   * Number.
-   */
-  template <typename OtherNumber>
-  Tensor (const Tensor<rank_,dim,OtherNumber> &initializer);
-
-  /**
-   * Constructor that converts from a "tensor of tensors".
-   */
-  template <typename OtherNumber>
-  Tensor (const Tensor<1,dim,Tensor<rank_-1,dim,OtherNumber> > &initializer);
-
-  /**
-   * Conversion operator to tensor of tensors.
-   */
-  template <typename OtherNumber>
-  operator Tensor<1,dim,Tensor<rank_-1,dim,OtherNumber> > () const;
-
-  /**
-   * Read-Write access operator.
-   */
-  Tensor<rank_-1,dim,Number> &operator [] (const unsigned int i);
-
-  /**
-   * Read-only access operator.
-   */
-  const Tensor<rank_-1,dim,Number> &operator [] (const unsigned int i) const;
-
-  /**
-   * Read access using TableIndices <tt>indices</tt>
-   */
-  Number operator [] (const TableIndices<rank_> &indices) const;
-
-  /**
-   * Read and write access using TableIndices <tt>indices</tt>
-   */
-  Number &operator [] (const TableIndices<rank_> &indices);
-
-  /**
-   * Copy assignment operator.
-   */
-  Tensor &operator = (const Tensor<rank_,dim,Number> &rhs);
-
-  /**
-   * Assignment operator from tensors with different underlying scalar type.
-   * This obviously requires that the @p OtherNumber type is convertible to @p
-   * Number.
-   */
-  template <typename OtherNumber>
-  Tensor &operator = (const Tensor<rank_,dim,OtherNumber> &rhs);
-
-  /**
-   * This operator assigns a scalar to a tensor. To avoid confusion with what
-   * exactly it means to assign a scalar value to a tensor, zero is the only
-   * value allowed for <tt>d</tt>, allowing the intuitive notation
-   * <tt>t=0</tt> to reset all elements of the tensor to zero.
-   *
-   * @relates EnableIfScalar
-   */
-  template <typename OtherNumber,
-            typename = typename EnableIfScalar<OtherNumber>::type>
-  Tensor<rank_,dim,Number> &operator = (const OtherNumber d);
-
-  /**
-   * Test for equality of two tensors.
-   */
-  template <typename OtherNumber>
-  bool operator == (const Tensor<rank_,dim,OtherNumber> &) const;
-
-  /**
-   * Test for inequality of two tensors.
-   */
-  template <typename OtherNumber>
-  bool operator != (const Tensor<rank_,dim,OtherNumber> &) const;
-
-  /**
-   * Add another tensor.
-   */
-  template <typename OtherNumber>
-  Tensor<rank_,dim,Number> &operator += (const Tensor<rank_,dim,OtherNumber> &);
-
-  /**
-   * Subtract another tensor.
-   */
-  template <typename OtherNumber>
-  Tensor<rank_,dim,Number> &operator -= (const Tensor<rank_,dim,OtherNumber> &);
-
-  /**
-   * Scale the tensor by <tt>factor</tt>, i.e. multiply all components by
-   * <tt>factor</tt>.
-   */
-  template <typename OtherNumber>
-  Tensor<rank_,dim,Number> &operator *= (const OtherNumber factor);
-
-  /**
-   * Scale the vector by <tt>1/factor</tt>.
-   */
-  template <typename OtherNumber>
-  Tensor<rank_,dim,Number> &operator /= (const OtherNumber factor);
-
-  /**
-   * Unary minus operator. Negate all entries of a tensor.
-   */
-  Tensor<rank_,dim,Number>   operator - () const;
-
-  /**
-   * Return the Frobenius-norm of a tensor, i.e. the square root of the sum of
-   * squares of all entries.
-   */
-  real_type norm () const;
-
-  /**
-   * Return the square of the Frobenius-norm of a tensor, i.e. the sum of
-   * squares of all entries.
-   *
-   * This function mainly exists because it makes computing the norm simpler
-   * recursively, but may also be useful in other contexts.
-   */
-  real_type norm_square () const;
-
-  /**
-   * Fill a vector with all tensor elements.
-   *
-   * This function unrolls all tensor entries into a single, linearly numbered
-   * vector. As usual in C++, the rightmost index of the tensor marches
-   * fastest.
-   */
-  template <typename OtherNumber>
-  void unroll (Vector<OtherNumber> &result) const;
-
-  /**
-   * Returns an unrolled index in the range [0,dim^rank-1] for the element of
-   * the tensor indexed by the argument to the function.
-   */
-  static
-  unsigned int
-  component_to_unrolled_index(const TableIndices<rank_> &indices);
-
-  /**
-   * Opposite of  component_to_unrolled_index: For an index in the range
-   * [0,dim^rank-1], return which set of indices it would correspond to.
-   */
-  static
-  TableIndices<rank_> unrolled_to_component_indices(const unsigned int i);
-
-  /**
-   * Reset all values to zero.
-   *
-   * Note that this is partly inconsistent with the semantics of the @p
-   * clear() member functions of the standard library containers and of
-   * several other classes within deal.II, which not only reset the values of
-   * stored elements to zero, but release all memory and return the object
-   * into a virginial state. However, since the size of objects of the present
-   * type is determined by its template parameters, resizing is not an option,
-   * and indeed the state where all elements have a zero value is the state
-   * right after construction of such an object.
-   */
-  void clear ();
-
-  /**
-   * Determine an estimate for the memory consumption (in bytes) of this
-   * object.
-   */
-  static std::size_t memory_consumption ();
-
-  /**
-   * Exception.
-   */
-  DeclException1 (ExcInvalidTensorContractionIndex,
-                  int,
-                  << "You have requested contraction of tensors over index "
-                  << arg1
-                  << ", but this is not possible for tensors of the current type.");
-
-  /**
-   * Read or write the data of this object to or from a stream for the purpose
-   * of serialization
-   */
-  template <class Archive>
-  void serialize(Archive &ar, const unsigned int version);
-
-private:
-  /**
-   * Array of tensors holding the subelements.
-   */
-  Tensor<rank_-1,dim,Number> subtensor[dim];
-
-  /**
-   * Help function for unroll.
-   */
-  template <typename OtherNumber>
-  void unroll_recursion(Vector<OtherNumber> &result,
-                        unsigned int    &start_index) const;
-
-  // make the following class a
-  // friend to this class. in principle,
-  // it would suffice if otherrank==rank+1,
-  // but then the compiler complains
-  // that this be an explicit specialization
-  // which is not what we want
-  //
-  // also, it would be sufficient to make
-  // the function unroll_loops a friend,
-  // but that seems to be impossible as well.
-  template <int, int, typename> friend class Tensor;
-};
-
-
-/*--------------------------- Inline functions -----------------------------*/
-
-#ifndef DOXYGEN
-
-template <int rank_, int dim, typename Number>
-inline
-Tensor<rank_,dim,Number>::Tensor (const bool initialize)
-{
-  if (initialize)
-    // need to create an object Number() to initialize to zero to avoid
-    // confusion with Tensor::operator=(scalar) when using something like
-    // Tensor<1,dim,Tensor<1,dim,Number> >.
-    for (unsigned int i=0; i!=dim; ++i)
-      subtensor[i] = Tensor<rank_-1,dim,Number>();
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-Tensor<rank_,dim,Number>::Tensor (const Tensor<rank_,dim,Number> &initializer)
-{
-  for (unsigned int i=0; i!=dim; ++i)
-    subtensor[i] = initializer[i];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-Tensor<rank_,dim,Number>::Tensor (const array_type &initializer)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] = initializer[i];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number>::Tensor (const Tensor<rank_,dim,OtherNumber> &initializer)
-{
-  for (unsigned int i=0; i!=dim; ++i)
-    subtensor[i] = initializer[i];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number>::Tensor
-(const Tensor<1,dim,Tensor<rank_-1,dim,OtherNumber> > &initializer)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] = initializer[i];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number>::operator
-Tensor<1,dim,Tensor<rank_-1,dim,OtherNumber> > () const
-{
-  return Tensor<1,dim,Tensor<rank_-1,dim,Number> > (subtensor);
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-typename Tensor<rank_,dim,Number>::value_type &
-Tensor<rank_,dim,Number>::operator[] (const unsigned int i)
-{
-  Assert (i<dim, ExcIndexRange(i, 0, dim));
-  return subtensor[i];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-const typename Tensor<rank_,dim,Number>::value_type &
-Tensor<rank_,dim,Number>::operator[] (const unsigned int i) const
-{
-  Assert (i<dim, ExcIndexRange(i, 0, dim));
-
-  return subtensor[i];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-Number
-Tensor<rank_,dim,Number>::operator[] (const TableIndices<rank_> &indices) const
-{
-  const unsigned int inner_ind = indices[0];
-  Assert (inner_ind<dim, ExcIndexRange(inner_ind, 0, dim));
-
-  TableIndices<rank_-1> indices1;
-  for (unsigned int i = 0; i < rank_-1; i++)
-    indices1[i] = indices[i+1];
-  return (subtensor[inner_ind])[indices1];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-Number &
-Tensor<rank_,dim,Number>::operator[] (const TableIndices<rank_> &indices)
-{
-  const unsigned int inner_ind = indices[0];
-  Assert (inner_ind<dim, ExcIndexRange(inner_ind, 0, dim));
-
-  TableIndices<rank_-1> indices1;
-  for (unsigned int i = 0; i < rank_-1; i++)
-    indices1[i] = indices[i+1];
-  return (subtensor[inner_ind])[indices1];
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator = (const Tensor<rank_,dim,Number> &t)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] = t.subtensor[i];
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator = (const Tensor<rank_,dim,OtherNumber> &t)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] = t.subtensor[i];
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber, typename>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator = (const OtherNumber d)
-{
-  Assert (d == OtherNumber(), ExcMessage ("Only assignment with zero is allowed"));
-  (void) d;
-
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] = Number();
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-bool
-Tensor<rank_,dim,Number>::operator == (const Tensor<rank_,dim,OtherNumber> &p) const
-{
-  for (unsigned int i=0; i<dim; ++i)
-    if (subtensor[i] != p.subtensor[i])
-      return false;
-  return true;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-bool
-Tensor<rank_,dim,Number>::operator != (const Tensor<rank_,dim,OtherNumber> &p) const
-{
-  return !((*this) == p);
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator += (const Tensor<rank_,dim,OtherNumber> &p)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] += p.subtensor[i];
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator -= (const Tensor<rank_,dim,OtherNumber> &p)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] -= p.subtensor[i];
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator *= (const OtherNumber s)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] *= s;
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-Tensor<rank_,dim,Number> &
-Tensor<rank_,dim,Number>::operator /= (const OtherNumber s)
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i] /= s;
-  return *this;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-Tensor<rank_,dim,Number>
-Tensor<rank_,dim,Number>::operator - () const
-{
-  Tensor<rank_,dim,Number> tmp;
-
-  for (unsigned int i=0; i<dim; ++i)
-    tmp.subtensor[i] = -subtensor[i];
-
-  return tmp;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-typename Tensor<rank_,dim,Number>::real_type
-Tensor<rank_,dim,Number>::norm () const
-{
-  return std::sqrt (norm_square());
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-typename Tensor<rank_,dim,Number>::real_type
-Tensor<rank_,dim,Number>::norm_square () const
-{
-  real_type s = 0;
-  for (unsigned int i=0; i<dim; ++i)
-    s += subtensor[i].norm_square();
-
-  return s;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-void
-Tensor<rank_, dim, Number>::unroll (Vector<OtherNumber> &result) const
-{
-  AssertDimension (result.size(),(Utilities::fixed_power<rank_, unsigned int>(dim)));
-
-  unsigned int index = 0;
-  unroll_recursion (result, index);
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <typename OtherNumber>
-inline
-void
-Tensor<rank_, dim, Number>::unroll_recursion (Vector<OtherNumber> &result,
-                                              unsigned int    &index) const
-{
-  for (unsigned int i=0; i<dim; ++i)
-    {
-      operator[](i).unroll_recursion(result, index);
-    }
-}
-
-template <int rank_, int dim, typename Number>
-inline
-unsigned int
-Tensor<rank_, dim, Number>::component_to_unrolled_index(const TableIndices<rank_> &indices)
-{
-  TableIndices<rank_-1> indices1;
-  for (unsigned int i = 0; i < rank_-1; i++)
-    indices1[i] = indices[i];
-
-  Assert (indices[rank_-1] < dim,
-          ExcIndexRange (indices[rank_-1], 0, dim));
-  return ( Tensor<rank_-1,dim,Number>::component_to_unrolled_index(indices1) * dim + indices[rank_-1]);
-}
-
-template <int rank_, int dim, typename Number>
-inline
-TableIndices<rank_>
-Tensor<rank_, dim, Number>::unrolled_to_component_indices(const unsigned int i)
-{
-  Assert (i < n_independent_components,
-          ExcIndexRange (i, 0, n_independent_components));
-
-  TableIndices<rank_>   indices;
-
-  unsigned int remainder = i;
-  for (int r=rank_-1; r>=0; --r)
-    {
-      indices[r] = (remainder % dim);
-      remainder /= dim;
-    }
-  Assert (remainder == 0, ExcInternalError());
-
-  return indices;
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-void Tensor<rank_,dim,Number>::clear ()
-{
-  for (unsigned int i=0; i<dim; ++i)
-    subtensor[i].clear();
-}
-
-
-
-template <int rank_, int dim, typename Number>
-inline
-std::size_t
-Tensor<rank_,dim,Number>::memory_consumption ()
-{
-  return sizeof(Tensor<rank_,dim,Number>);
-}
-
-
-
-template <int rank_, int dim, typename Number>
-template <class Archive>
-inline
-void
-Tensor<rank_,dim,Number>::serialize(Archive &ar, const unsigned int)
-{
-  ar &subtensor;
-}
-
-#endif // DOXYGEN
 /* ----------------- Non-member functions operating on tensors. ------------ */
